+++ src/config.h
@@ -66,13 +66,15 @@ struct ConfigSettingsStruct{
   char webPass[50];
   bool disableLedBlue;
   bool disableLedYellow;
-  bool disablePingCtrl;
+  //bool disablePingCtrl;
   bool disableLeds;
   COORDINATOR_MODE_t coordinator_mode;
   COORDINATOR_MODE_t prevCoordinator_mode;//for button
   bool keepWeb;
   bool apStarted;
   bool wifiWebSetupInProgress;
+  bool fwEnabled;
+  IPAddress fwIp;
 };
 
 struct InfosStruct